requirements.txt
# Core Dependencies
pytest==7.4.3
pytest-playwright==0.4.3
playwright==1.40.0
requests==2.31.0
pyyaml==6.0.1
python-dotenv==1.0.0
faker==20.1.0

# Reporting
allure-pytest==2.13.2
pytest-html==4.1.1

# Parallel Execution
pytest-xdist==3.5.0

# Code Quality (optional, for development)
black==23.12.1
flake8==6.1.0

# Utilities
python-dateutil==2.8.2
jinja2==3.1.2
pillow==10.1.0

# Validation
jsonschema==4.20.0
deepdiff==6.7.1